+++ src/bin/df_test.rs
@@ -4,12 +4,13 @@ use self::Header::*;
 use std::fmt;
 
 fn main() {
-    let a = DataFrame::new(vec![X, Y], vec![c!(1,2,3), c!(4,5,6)]);
+    let a = DataFrame::new(vec![X, Y, Z], vec![c!(1,2,3), c!(4,5,6), c!(7,8,9)]);
     println!("{}", a);
 }
 
 #[derive(Debug, Hash, PartialEq, Eq, Clone)]
 pub enum Header {
     X,
-    Y
+    Y,
+    Z
 }